% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_report_templates.R
\name{available_report_templates}
\alias{available_report_templates}
\title{Get Available Report Templates}
\usage{
available_report_templates()
}
\value{
\code{character()} Vector of template names
}
\description{
Get the names of the valid templates for use when generating reports
}
\examples{
available_report_templates()
}
