% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_add}
\alias{example_add}
\title{Example Addition Function Title}
\usage{
example_add(a, b)
}
\arguments{
\item{a}{\code{numeric()} First number to add}

\item{b}{\code{numeric()} Second number to add}
}
\value{
The addition of \code{a} and \code{b}
}
\description{
Description of example addition function. Here we put a short paragraph
describing what the function does at a high level.
}
\examples{
example_add(1, 2)
example_add(5.4, 3.2)
example_add(3.3, 1:10)

\dontrun{
example_add("a", "b")
}
}
