% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expected_level}
\alias{expected_level}
\title{The expected level analysis}
\format{
A data frame with 9 rows and 3 variables:
\describe{
\item{old_outcome}{The original outcome of interest}
\item{new_outcome}{The revised outcome of interest}
\item{year}{Year the data was "collected"}
\item{group}{The group associated with the outcome}
\item{pert_diff}{The percentage difference between new_outcome and
old_outcome}
\item{max_abs_pert_diff}{The maximum absolute percentage difference in
each group}
\item{mean_abs_pert_diff}{The mean absolute percentage difference in each
group}
\item{rank}{The ranking of the differences in the outcome variable between
the new and old data where 1 is the greatest mean_abs_pert_diff.}
}
}
\usage{
expected_level
}
\description{
The expected level analysis associated with the test on old_data and
new_data by the grouping variable. The variables are as follows:
}
\keyword{datasets}
