% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate Diagnostic Report}
\usage{
generate_report(template, params = NULL, output_dir = NULL, ...)
}
\arguments{
\item{template}{\code{character()} Name of template used to generate report.}

\item{params}{\code{list()} Parameters passed along to the chosen report template.}

\item{output_dir}{\code{character()} Directory to write generated report to.
Defaults to the current working directory.}

\item{...}{Additional arguments passed on to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\value{
\code{character()} Invisible path to the generated report.
}
\description{
Using one of the included templates, generate a diagnostic report based on a
set of input parameters.
}
\examples{
\dontrun{
# Proof of concept report
generate_report("test_report")

# Basic tend/level summary report
report_params <- list(
  old_data = old_data,
  new_data = new_data,
  level_threshold = 20,
  trend_bins = 1,
  comparison_var = "outcome",
  id_vars = c("year", "group"),
  group_vars = "group",
  trend_var = "year",
  scatter_color_var = "group"
)

generate_report(
  "level_trend_summary",
  params = report_params,
  output_dir = "~"
)

}
}
