% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{new_data_alt}
\alias{new_data_alt}
\title{Another outcome variable measured over time for different groups. Measured
using a different method and/or different input data.}
\format{
A data frame with 6 rows and 3 variables:
\describe{
\item{outcome}{An outcome of interest}
\item{year}{Year the data was "collected"}
\item{group}{The group associated with the outcome}
}
}
\usage{
new_data_alt
}
\description{
An example data set with an outcome and a grouping variable. The variables
are as follows:
}
\keyword{datasets}
