% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{old_data}
\alias{old_data}
\title{An old outcome variable measured over time for different groups}
\format{
A data frame with 9 rows and 3 variables:
\describe{
\item{outcome}{An outcome of interest}
\item{year}{Year the data was "collected"}
\item{group}{The group associated with the outcome}
}
}
\usage{
old_data
}
\description{
An example data set with an outcome and a grouping variable. The variables
are as follows:
}
\keyword{datasets}
