% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{plot_scatter}
\usage{
plot_scatter(
  data,
  x_var,
  y_var,
  facet_x,
  facet_y = NULL,
  color_pt = NULL,
  pt_size = 1,
  title = NULL,
  facet_type = "none",
  fix_scale = FALSE
)
}
\arguments{
\item{data}{data.frame or data.table. Data to plot comparing two sets of estimates, e.g.
output of rank_change function. Required!}

\item{x_var}{\code{character()} Variable to plot on the x-axis, e.g the "old" version or
previous estimate. Required!}

\item{y_var}{\code{character()} Variable to plot on the y-axis, e.g. the "new" version or
updated estimate. Required!}

\item{facet_x}{\code{character()} Variable to facet by, along x-axis. This is required in order
to use facet_wrap and facet_grid. Defaults to NULL.}

\item{facet_y}{\code{character()} Variable to facet by, along y-axis, if desired. This is required
in order to use facet_grid. Defaults to NULL.}

\item{color_pt}{\code{character()} Variable to modify color of geom_point, if desired.
Defaults to NULL.}

\item{pt_size}{\code{numeric()} Option to modify point size. Defaults to 1.}

\item{title}{\code{character()} Title for plot, if desired. Defaults to NULL.}

\item{facet_type}{\code{character()} Choice of "none", "wrap", or "grid". Defaults to "none". The
"none" option will create a single plot. The "wrap" option will produce a plot using facet_wrap.
The "grid" option will produce a plot using facet_grid.}

\item{fix_scale}{\code{boolean()} Option to fix the axes for plot, using the min and max values.
Defaults to FALSE. Will override scales = "free".}
}
\value{
A ggplot object
}
\description{
This function can be used to visualize the results of the rank_change function, or to otherwise
compare values from two different model versions or sets of estimates. This can help streamline
the vetting process. All characteristics for each estimate should be matched, e.g. year,
location, age. Plot_scatter creates scatter plots comparing the two sets of results. The
various options for faceting or coloring the points allow the user to customize the plot to
maximize utility.
}
\examples{
plot_scatter(trend_changes_USA, 
             x_var = "old_mean", 
             y_var = "new_mean", 
             facet_x = "ihme_loc_id", 
             pt_size = 2, 
             facet_type = "wrap", 
             fix_scale = TRUE, 
             color_pt = "year_id")
}
