% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_change.R
\name{rank_changes}
\alias{rank_changes}
\title{rank_change}
\usage{
rank_changes(
  new_data,
  old_data,
  change_type,
  threshold = 1,
  comparison_var,
  id_vars,
  group_vars,
  trend_var = NULL
)
}
\arguments{
\item{new_data}{\code{data.table()} New version of data for comparison.}

\item{old_data}{\code{data.table()} Old version of data for comparison.}

\item{change_type}{\code{character()} Type of comparison to be made. Must be one
of "level" or "trend".}

\item{threshold}{\code{numeric()} The value of threshold depends on
change_type. For level changes, threshold must be between 0 and 100
(inclusive). It represents the maximum absolute percent difference of
comparison_var by group_vars to be included in the analysis. For trend
changes, threshold represents the number of bins to use in the trend
analysis within each grouping by group_vars. The default threshold is
1, which would examine the change in the slope (i.e. change in comparison_var
over change in trend_var) by group_vars in the trend analysis, and would
examine changes greater than 1\% by group_vars in the level analysis.}

\item{comparison_var}{\code{numeric()} Outcome variable to make comparisons
over.}

\item{id_vars}{Vector of variable names that uniquely identify
both new_data and old_data.}

\item{group_vars}{Level to summarize change variables for
ranking. group_vars are a subset of id_vars.}

\item{trend_var}{\code{numeric()} Predictor variable used to examine trend
changes. trend_var is a subset of id_vars of length 1.}
}
\value{
\code{data.table()} of ranked merged data.tables, which may be a subset
of the input data.tables. Rank number 1 corresponds to the greatest change
in comparison_var by group_vars between new_data and old_data, rank
number 2 corresponds to the second greatest change in comparison_var by
group_vars, etc. NOTE: Ranks may be duplicated by group if they have the same
values used for ranking.
}
\description{
Make comparisons between two versions of the same outcome by some grouping
variables. rank_change can make level (i.e. percentage change) and
trend (i.e. linear trend) comparisons. Level changes are ranked by the
greatest mean absolute percentage change to the smallest. Trend changes are
ranked first by whether or not the slope changed sign and then by the
greatest mean absolute percentage change to the smallest.
}
\examples{
\dontrun{
rank_change(new_data,
            old_data,
            change_type = "level",
            threshold = 20,
            comparison_var = "outcome",
            id_vars = c("sex", "country", "year"),
            group_vars = c("sex", "country"),
            )

rank_change(new_data,
            old_data,
            change_type = "trend",
            threshold = 3,
            comparison_var = "outcome",
            id_vars = c("sex", "country", "year"),
            group_vars = c("sex", "country"),
            trend_var = "year"
            )
}
}
