% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_level.R
\name{rank_level}
\alias{rank_level}
\title{rank_level}
\usage{
rank_level(
  new_data,
  old_data,
  threshold = 0,
  comparison_var,
  id_vars,
  group_vars,
  displace_old_zero = 1e-10
)
}
\arguments{
\item{new_data}{\code{data.table()} New version of data for comparison.}

\item{old_data}{\code{data.table()} Old version of data for comparison.}

\item{threshold}{\code{numeric()} The threshold must be greater than or equal
to zero. It represents the maximum absolute percent difference of
comparison_var by group_vars to be included in the analysis. The default is
zero.}

\item{comparison_var}{\code{numeric()} Outcome variable to make comparisons
over.}

\item{id_vars}{Vector of variable names that uniquely identify
both new_data and old_data.}

\item{group_vars}{Level to summarize change variables for
ranking. group_vars are a subset of id_vars.}

\item{displace_old_zero}{\code{numeric()} Displace zeros in old_data to prevent
NaNs in change calculations. Default is 1e-10.}
}
\value{
\code{data.table()} of ranked merged data.tables, which may be a subset
of the input data.tables. Rank number 1 corresponds to the greatest change
in comparison_var by group_vars between new_data and old_data, rank
number 2 corresponds to the second greatest change in comparison_var by
group_vars, etc. NOTE: Ranks may be duplicated by group if they have the same
values used for ranking.
}
\description{
Make comparisons between two versions of the same outcome by some grouping
variables and rank them by the greatest mean absolute percentage
change to the smallest.
}
\examples{
rank_level(new_data,
            old_data,
            threshold = 5,
            comparison_var = "outcome",
            id_vars = c("year", "group"),
            group_vars = "group")
}
