% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate Diagnostic Report}
\usage{
generate_report(template, params = NULL, output_dir = NULL, ...)
}
\arguments{
\item{template}{\code{character()} Name of template used to generate report.}

\item{params}{\code{list()} Parameters passed along to the chosen report template.}

\item{output_dir}{\code{character()} Directory to write generated report to.
Defaults to the current working directory.}

\item{...}{Additional arguments passed on to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\value{
\code{character()} Invisible path to the generated report.
}
\description{
Using one of the included templates, generate a diagnostic report based on a
set of input parameters.
}
\examples{
\dontrun{
generate_report("test_report")
}
}
